Added footstep sounds. Fixed some errors

+++ game/DialogueHandler.gd
@@ -17,6 +17,8 @@ func _ready() -> void:
 
 
 func show_dialogue(dialogue:PoolStringArray, _dialogue_speed:float = 0.05) -> void:
+	if dialogue[0].begins_with("tip") && (!text_timer.is_stopped() || !full_dialogue.empty()):
+		return
 	self.dialogue_speed = _dialogue_speed
 	fade_tween.remove_all()
 	delay_timer.stop()
